+++ air/src/errors.rs
@@ -1,5 +1,3 @@
-use crate::trace::MIN_TRACE_LEN;
-
 // EXECUTION OPTIONS ERROR
 // ================================================================================================
 
@@ -9,8 +7,8 @@ pub enum ExecutionOptionsError {
         "expected number of cycles {expected_cycles} must be smaller than the maximum number of cycles {max_cycles}"
     )]
     ExpectedCyclesTooBig { max_cycles: u32, expected_cycles: u32 },
-    #[error(
-        "maximum number of cycles {0} must be greater than the minimum number of cycles {MIN_TRACE_LEN}"
-    )]
-    MaxCycleNumTooSmall(u32),
+    #[error("maximum number of cycles {max_cycles} must be greater than {min_cycles_limit}")]
+    MaxCycleNumTooSmall { max_cycles: u32, min_cycles_limit: usize },
+    #[error("maximum number of cycles {max_cycles} must be less than {max_cycles_limit}")]
+    MaxCycleNumTooBig { max_cycles: u32, max_cycles_limit: u32 },
 }